First release.

+++ windows-kill-library/windows-kill-library.cpp
@@ -5,8 +5,8 @@
 
 namespace WindowsKillLibrary
 {
-	void sendSignal(int signal_type, int signal_pid) {
-		Signal the_signal((DWORD)signal_type, (DWORD)signal_pid);
+	void sendSignal(DWORD signal_type, DWORD signal_pid) {
+		Signal the_signal(signal_type, signal_pid);
 		Sender::send(the_signal);
 	}
 };
